Rename to vqvae
