+++ src/torchsmith/models/vae/vqvae.py
