+++ tests/models/test_vae_utils.py
@@ -9,7 +9,7 @@ from torchsmith.models.vae.base import BaseVAE
 from torchsmith.models.vae.utils import generate_reconstructions
 from torchsmith.models.vae.utils import generate_samples
 from torchsmith.models.vae.vae_conv import VAEConv
-from torchsmith.models.vae.vq_vae import VQVAE
+from torchsmith.models.vae.vqvae import VQVAE
 from torchsmith.utils.constants import RANDOM_STATE
 from torchsmith.utils.plotting import suppress_plot
 from torchsmith.utils.pytorch import get_device
