+++ tests/training/test_trainer_vqvae_autoregression.py
@@ -7,7 +7,7 @@ from torchsmith.datahub.images_with_vqvae import ImagesWithVQVAEDataset
 from torchsmith.datahub.svhn import postprocess_data
 from torchsmith.datahub.svhn import preprocess_data
 from torchsmith.models.gpt2 import GPT2Decoder
-from torchsmith.models.vae.vq_vae import VQVAE
+from torchsmith.models.vae.vqvae import VQVAE
 from torchsmith.tokenizers.vqvae_tokenizer import VQVAEImageTokenizer
 from torchsmith.tokenizers.vqvae_tokenizer import generate_samples_image_v2
 from torchsmith.training.config import GPT2Config
